fatigue bugs

+++ drf_sp_hub/sp_api/signals.py
@@ -81,6 +81,6 @@ def delete_html_file_on_change(sender, instance, **kwargs):
         return False
 
     new_file = instance.html_file
-    if not old_file == new_file:
+    if old_file and not old_file == new_file:
         if os.path.isfile(old_file.path):
             os.remove(old_file.path)
